made a game over and a check board function for tic tac toe.

+++ src/examples/04_module/01_bank/checking_account.h
@@ -6,5 +6,5 @@ class CheckingAccount : public BankAccount
 public:
 	CheckingAccount() = default;
 	explicit CheckingAccount(int b) : BankAccount(b) {}
-	int get_balance()const { return balance * (1 + get_rate()); }
+	double get_balance()const { return balance * (1 + get_rate()); }
 };
